fix(shared): Update GitHub PAT, closes #5820
This was generating an error because the poersonal access token for
GitHub had expired (the one we use to pull a list of known issues from
the GitHub Graph API.

I have regerenated the token and udated it.
Furthermore, I did not fix the error, because I am 100% certain going to
forget about renewing this token 1 year from now, and at least this way
it will bubble to the surface.