[shared] fix: Pattern translation in UI
This fixes problems with pattern translation where the translation
inside a pattern we using the UI's i18n instance to do the translation.

That causes issues when the pattern language and UI langauges are
different (something we support). Furthermore, we need to load the
design's own namespace as the default, and load the annotations plugin
namespace and do some other things that are specific to translation
inside the patterns.

This resolves that by creating a dedicated i18n instance to handle
translation inside patterns.

This unblocks #6340