chore(core): Drop the chai-string dependency for tests
We're using chai-string as it provides the `equalIgnoreSpaces`
assertion which we use in our SVG tests.

However, that package does not seem to be maintained, and lists chai v4
as its peer dependency. We've moved on to chai v5 and this is causes
issues in one of our github workflows as dependencies cannot be
resolved.

So, I've extracted the assertion we need, and dropped the dependency.