fix(huey): Zipper length calculation, clarify it is a maximum
This stems from the excellent work by @HaasJona in #6421 in which they
requested to double-check the zipper length calculation. Which I didn't
get around to doing in the code review, but was able to now.

Turns out there was a slight issue with it, so this resolves it.
I've also updated the wording to clarify that what we are calculating as
the __maximum__ zipper lenght. Since zippers typically come is standard
lengths, it's typically not possible to find the exact length you need,
so it's IMHO important to make it clear to the user they should always
purchase a zipper that is as close as possible, yet shorter than the
calculated maximum length.