wip: Migrating snippets from defs to symbols
@roel-v requested in #6359 whether we could use symbol instead of defs
to place snippets on the pattern with the use tag.

I'm not against that as such, but a symbol and a def are different
things, and an initial test (this commit) shows that there's likely to
be a bunch of knock-on effects.

Symbols have their own viewBox, but ok, we can work with that.
I've gotten it to the point where things look ok in FreeSewing, but if
you open the generated SVG in Inkscape or another pattern editor, the
symbols are all over the place.

So, I took a stab at this, but this seems to open a can of worms for
rather little gain.

Still, I'm going to report back to @roel-v as perhaps they have
arguments or ideas that could sway my opinion.